==> 1_data_preprocessing.ipynb <==
# Data Preprocessing Tools

## Importing the libraries

import os
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd

## Importing the dataset

# Load excel file
Output_Precovid_df = pd.read_excel(r"C:\Users\LakshmanNavaneethakr\Downloads\lakshmansierra-git\hotel_review_regression\00-raw-data\Output_Precovid.xlsx")
postcovid_output_df = pd.read_excel(r"C:\Users\LakshmanNavaneethakr\Downloads\lakshmansierra-git\hotel_review_regression\00-raw-data\postcovid_output.xlsx")

os.makedirs("00-raw-csv-data", exist_ok=True)
Output_Precovid_df.to_csv("00-raw-csv-data/Output_Precovid.csv", index=None, header=True)
postcovid_output_df.to_csv("00-raw-csv-data/postcovid_output.csv", index=None, header=True)

# Load csv file
Output_Precovid_df = pd.read_csv(r"C:\Users\LakshmanNavaneethakr\Downloads\lakshmansierra-git\hotel_review_regression\00-raw-csv-data\Output_Precovid.csv")
print(Output_Precovid_df.tail())

# Load csv file
postcovid_output_df = pd.read_csv(r"C:\Users\LakshmanNavaneethakr\Downloads\lakshmansierra-git\hotel_review_regression\00-raw-csv-data\postcovid_output.csv")
print(postcovid_output_df.tail())

print(Output_Precovid_df.shape)
print(postcovid_output_df.shape)

print(Output_Precovid_df.columns)
print(postcovid_output_df.columns)

Output_Precovid_df_columns = list(Output_Precovid_df.columns)
print(type(Output_Precovid_df_columns))

postcovid_output_df_columns = list(postcovid_output_df.columns)
print(type(postcovid_output_df_columns))

x = []
for col in Output_Precovid_df_columns:
    if col in postcovid_output_df_columns:
        x.append(col)
        
    else:
        print("missing")
        
print(len(x))

Output_Precovid_df_dummies = Output_Precovid_df.copy()
postcovid_output_df_dummies = postcovid_output_df.copy()
Output_Precovid_df_dummies['dummies'] = 0
postcovid_output_df_dummies['dummies'] = 1

Output_Precovid_df_dummies.columns
postcovid_output_df_dummies.columns

print(Output_Precovid_df_dummies.head())
print(postcovid_output_df_dummies.head())

## Concat

concat_df = pd.concat([Output_Precovid_df_dummies, postcovid_output_df_dummies])

print(concat_df.head())
print(concat_df.tail())
print(concat_df.shape)


## Saving csv

os.makedirs("01-concat_data", exist_ok=True)
concat_df.to_csv("01-concat_data/pre_post_covid.csv", index=False)
