==> 2_data_preprocessing.ipynb <==
# Data Preprocessing Tools

## Importing the libraries

import os
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd

## Importing the dataset

# Load csv file
concat_df = pd.read_csv(r"C:\Users\LakshmanNavaneethakr\Downloads\lakshmansierra-git\hotel_review_regression\01-concat_data\pre_post_covid.csv", encoding="latin1")

print(concat_df.shape)

print(concat_df.columns)

## Encode month of occurence

concat_df['Date of stay'] = pd.to_datetime(concat_df['Date of stay'], errors='coerce')
concat_df['Month'] = concat_df['Date of stay'].dt.month

month_dummies = pd.get_dummies(concat_df['Month'], prefix='M', prefix_sep='')

for i in range(1, 13):
    col = f'M{i}'
    if col not in month_dummies.columns:
        month_dummies[col] = 0

month_dummies = month_dummies.reindex(sorted(month_dummies.columns, key=lambda x: int(x[1:])), axis=1)

month_dummies = month_dummies.astype(int)

concat_month_dummies_df = pd.concat([concat_df, month_dummies], axis=1)
concat_month_dummies_df.drop(columns=['Month', 'M1'], inplace=True)


print(concat_month_dummies_df.columns)

print(concat_month_dummies_df.shape)

## Add Review Description length column 

concat_month_dummies_df['review_length'] = (
    concat_month_dummies_df['Review Description']
    .fillna('')                            # Replace NaN with empty string
    .astype(str)                           # Ensure all are strings
    .str.strip()                           # Remove spaces before/after
    .apply(lambda x: len(x) if x else 0)   # Length = len(x) else 0 if empty
)


print(concat_month_dummies_df.columns)

print(concat_month_dummies_df.shape)

## Saving csv

os.makedirs("02-encoded_month", exist_ok=True)
concat_month_dummies_df.to_csv("02-encoded_month/pre_post_covid_month_encoded.csv", index=False)

os.makedirs("02-encoded_month", exist_ok=True)
concat_month_dummies_df.head(50).to_csv("02-encoded_month/pre_post_covid_month_encoded_head.csv", index=False)
concat_month_dummies_df.tail(50).to_csv("02-encoded_month/pre_post_covid_month_encoded_tail.csv", index=False)
