==> 4_statsmodel.ipynb <==
# Data Preprocessing Tools

## Importing the libraries

import pandas as pd
import numpy as np
import statsmodels.api as sm
from scipy import stats
from statsmodels.stats.anova import anova_lm
import os

## Importing the dataset

df = pd.read_csv(r"C:\Users\LakshmanNavaneethakr\Downloads\lakshmansierra-git\hotel_review_regression\02-encoded_month\pre_post_covid_month_encoded.csv", encoding="latin1")

print(df.shape)

print(df.columns)

## Selective columns

selected_columns = [
    'dummies',
    'Total_aggregation_for_match_stay',
    'Total_aggregation_for_match_food',
    'Total_aggregation_for_match_clean',
    'Total_aggregation_for_match_location',
    'Total_aggregation_for_match_cost',
    'Total_aggregation_for_match_safety',
    'Total_aggregation_for_match_service',
    'Total_aggregation_for_match_decoration',
    'Total_aggregation_for_match_facility',
    'Total_aggregation_for_match_ambience',
    'Total_aggregation_for_match_room',
    'M2', 
    'M3', 
    'M4', 
    'M5', 
    'M6', 
    'M7',
    'M8', 
    'M9', 
    'M10', 
    'M11', 
    'M12',
    'review_length',
    'Review_Score'
]

selected_df = df[selected_columns].copy()


print(selected_df.columns)

print(selected_df.shape)

print(selected_df.dtypes)

## filter post covid data with `dummies = 1`

post_covid_df = selected_df[selected_df['dummies'] == 1].copy()


print(post_covid_df.shape)

## drop dummies

post_covid_ml_df = post_covid_df.copy()

post_covid_ml_df.drop(columns=['dummies'], inplace=True)

print(post_covid_ml_df.shape)

post_covid_ml_df.head()

## Loading X and Y

X = post_covid_ml_df.iloc[:, :-1].values
y = post_covid_ml_df.iloc[:, -1].values

print(X.shape)
print(y.shape)

## Observation

# Add constant (intercept)
X_sm = sm.add_constant(X)

# Fit full regression model
model_full = sm.OLS(y, X_sm).fit()


### Metrics table

n = int(model_full.nobs)
p = X.shape[1]
f_stat = model_full.fvalue
f_pvalue = model_full.f_pvalue
r2 = model_full.rsquared
adj_r2 = model_full.rsquared_adj
root_mse = np.sqrt(model_full.mse_resid)

metrics_table_df = pd.DataFrame({
    'Metric': [
        'Number of obs',
        f'F( {p}, {n - p - 1} )',
        'Prob > F',
        'R-squared',
        'Adj R-squared',
        'Root MSE'
    ],
    'Value': [
        n,
        round(f_stat, 2),
        f'{f_pvalue:.4f}',
        round(r2, 4),
        round(adj_r2, 4),
        round(root_mse, 5)
    ]
})

#### Save metrics table

os.makedirs("06-metrics-table", exist_ok=True)
metrics_table_df.to_csv("06-metrics-table/06_metrics_table.csv", index=False)

### ANOVA table

import statsmodels.formula.api as smf

dependent_var = post_covid_ml_df.columns[-1]
independent_vars = " + ".join(post_covid_ml_df.columns[:-1])
formula = f"{dependent_var} ~ {independent_vars}"

model_full = smf.ols(formula=formula, data=post_covid_ml_df).fit()


#### Feature level ANOVA table

# anova_results = anova_lm(model_full)
# anova_table = anova_results.reset_index()
# anova_table.rename(columns={'index': 'Source'}, inplace=True)

# # Format like Stata
# anova_table['SS'] = anova_table['sum_sq'].map('{:.6f}'.format)
# anova_table['df'] = anova_table['df'].astype(int)
# anova_table['MS'] = (anova_results['mean_sq']).map('{:.8f}'.format)
# anova_table = anova_table[['Source', 'SS', 'df', 'MS']]

# # Add Total row manually
# total_ss = anova_results['sum_sq'].sum()
# total_df = int(anova_results['df'].sum())
# total_ms = total_ss / total_df
# anova_table.loc[len(anova_table)] = ['Total',
#     f'{total_ss:.6f}', total_df, f'{total_ms:.8f}']

#### Model level ANOVA table

anova_results = anova_lm(model_full, typ=1)

# Aggregate all predictors’ SS and df into a single “Model” entry
ss_model = anova_results.loc[anova_results.index != 'Residual', 'sum_sq'].sum()
df_model = anova_results.loc[anova_results.index != 'Residual', 'df'].sum()
ss_resid = anova_results.loc['Residual', 'sum_sq']
df_resid = anova_results.loc['Residual', 'df']
ss_total = ss_model + ss_resid
df_total = df_model + df_resid

# Compute Mean Squares (MS)
ms_model = ss_model / df_model
ms_resid = ss_resid / df_resid
ms_total = ss_total / df_total

# Build Stata-style table
anova_table = pd.DataFrame({
    'Source': ['Model', 'Residual', 'Total'],
    'SS': [f'{ss_model:.6f}', f'{ss_resid:.6f}', f'{ss_total:.6f}'],
    'df': [int(df_model), int(df_resid), int(df_total)],
    'MS': [f'{ms_model:.8f}', f'{ms_resid:.8f}', f'{ms_total:.8f}']
})


#### Save ANOVA table

os.makedirs("07-anova-table", exist_ok=True)
anova_table.to_csv("07-anova-table/07_anova_table.csv", index=False)

### Coefficient Table

params = model_full.params
bse = model_full.bse
tvalues = model_full.tvalues
pvalues = model_full.pvalues
conf = model_full.conf_int()

coeff_table_df = pd.DataFrame({
    'Variable': params.index,
    'Coef.': params.values,
    'Std. Err.': bse.values,
    't': tvalues.values,
    'P>|t|': pvalues.values,
    'Lower 95%': conf[0].values,
    'Upper 95%': conf[1].values
})

# Formatting like your image
coeff_table_df['Coef.'] = coeff_table_df['Coef.'].map('{:.7f}'.format)
coeff_table_df['Std. Err.'] = coeff_table_df['Std. Err.'].map('{:.7f}'.format)
coeff_table_df['t'] = coeff_table_df['t'].map('{:.2f}'.format)
coeff_table_df['P>|t|'] = coeff_table_df['P>|t|'].map('{:.3f}'.format)
coeff_table_df['Lower 95%'] = coeff_table_df['Lower 95%'].map('{:.6f}'.format)
coeff_table_df['Upper 95%'] = coeff_table_df['Upper 95%'].map('{:.6f}'.format)
coeff_table_df.loc[coeff_table_df['Variable'] == 'const', 'Variable'] = '_cons'
coeff_table_df.reset_index(drop=True, inplace=True)


#### Save coefficient table

os.makedirs("08-coeff-table", exist_ok=True)
coeff_table_df.to_csv("08-coeff-table/08_coeff_table.csv", index=False)
